% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{as.list.iteror}
\alias{as.list.iteror}
\alias{as.double.iteror}
\alias{as.numeric.iteror}
\alias{as.logical.iteror}
\alias{as.character.iteror}
\alias{as.vector.iteror}
\title{Collect all (or some given number of) values from an iteror,
returning a vector of the given type.}
\usage{
\method{as.list}{iteror}(x, n = as.integer(2^31 - 1), ...)

\method{as.double}{iteror}(x, n = as.integer(2^31 - 1), ...)

\method{as.numeric}{iteror}(x, n = as.integer(2^31 - 1), ...)

\method{as.logical}{iteror}(x, n = as.integer(2^31 - 1), ...)

\method{as.character}{iteror}(x, n = as.integer(2^31 - 1), ...)

\method{as.vector}{iteror}(x, mode)
}
\arguments{
\item{x}{an iterable object}

\item{n}{the maximum number of elements to return.}

\item{...}{Unused arguments will throw an error.}

\item{mode}{What mode to use for the output vector.}
}
\value{
The returned value will be \code{n} elements long if the
iterator did not stop.
}
\description{
Collect all (or some given number of) values from an iteror,
returning a vector of the given type.
}
\seealso{
concat take
}
