% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{chain}
\alias{chain}
\title{Create a chaining iterator}
\usage{
chain(...)
}
\arguments{
\item{\dots}{The iterables to iterate over.}
}
\description{
Create an iterator that chains multiple iterables together.
}
\details{
Originally from the `itertools` package.
}
\examples{

# Iterate over two iterables
as.list(chain(1:2, letters[1:3]))

}
\keyword{utilities}
