% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilength.R
\name{count}
\alias{count}
\title{Consumes an iterator and computes its length}
\usage{
count(object, ...)
}
\arguments{
\item{object}{an iterable object}

\item{...}{passed along to \link{iteror} constructor.}
}
\value{
the number of elements in the iterator
}
\description{
Counts the number of elements in an iterator. NOTE: The iterator is consumed
in the process.
}
\examples{
count(1:5) == length(1:5)

it <- iteror(1:5)
count(it) == length(1:5)

it2 <- i_chain(1:3, 4:5, 6)
count(it2)

it3 <- i_chain(1:3, levels(iris$Species))
count(it3)

}
\seealso{
take consume as.list.iteror
}
