% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotproduct.R
\name{dotproduct}
\alias{dotproduct}
\title{Computes the dot product of two iterable objects.}
\usage{
dotproduct(vec1, vec2)
}
\arguments{
\item{vec1}{the first}

\item{vec2}{the second iterable object}
}
\value{
the dot product of the iterators
}
\description{
Returns the dot product of two numeric iterables of equal length
}
\examples{
it <- iteror(1:3)
it2 <- iteror(4:6)
dotproduct(it, it2) # 32

it <- iteror(1:4)
it2 <- iteror(7:10)
dotproduct(1:4, 7:10) # 90

}
