% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerate.R
\name{enumerate}
\alias{enumerate}
\title{Create an enumeration object}
\usage{
enumerate(iterable)
}
\arguments{
\item{iterable}{Iterable to iterate over.}
}
\description{
Create an iterator that iterates over an iterable, returning the value in a
list that includes an index.
}
\examples{

# Create an enumeration of five random numbers
as.list(enumerate(rnorm(5)))

}
\keyword{utilities}
