% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteror.R
\name{hasNext}
\alias{hasNext}
\alias{hasNext.ihasNext}
\title{Does This Iterator Have A Next Element}
\usage{
hasNext(obj, ...)
}
\arguments{
\item{obj}{an iterator object.}

\item{\dots}{additional arguments that are ignored.}
}
\value{
Logical value indicating whether the iterator has a next element.
}
\description{
\code{hasNext} is a generic function that indicates if the iterator has
another element.
}
\examples{

  it <- ihasNext(iteror(c('a', 'b', 'c')))
  while (hasNext(it))
    print(nextOr(it))

}
\keyword{methods}
