% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilimit.R
\name{i_limit}
\alias{i_limit}
\title{Create a limited iterator}
\usage{
i_limit(iterable, n, ...)
}
\arguments{
\item{iterable}{Iterable to iterate over.}

\item{n}{Maximum number of values to return.}

\item{...}{Extra arguments for \code{iteror(iterable, ...)}}
}
\description{
Create an iterator that wraps a specified iterable a limited number of
times.
}
\details{
Originally from the \code{itertools} package.
}
\examples{

# Limit icount to only return three values
as.list(i_limit(icount(), 3))

}
\keyword{utilities}
