% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ireaddf.R
\name{i_readdf}
\alias{i_readdf}
\alias{ireaddf}
\title{Create an iterator to read data frames from files}
\usage{
i_readdf(filenames, n, start = 1, col.names, chunkSize = 1000)
}
\arguments{
\item{filenames}{Names of files contains column data.}

\item{n}{Maximum number of elements to read from each column file.}

\item{start}{Element to start reading from.}

\item{col.names}{Names of the columns.}

\item{chunkSize}{Number of rows to read at a time.}
}
\value{
An \link{iteror} yielding \link{data.frame} objects with up to \code{n} rows.
}
\description{
Create an iterator to read data frames from files.
}
\details{
Originally from the \code{itertools} package.
}
