% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ireadLines.R
\name{i_readLines}
\alias{i_readLines}
\alias{ireadLines}
\title{Iterator over Lines of Text from a Connection}
\usage{
i_readLines(con, n = 1, ...)
}
\arguments{
\item{con}{a connection object or a character string.}

\item{n}{integer.  The maximum number of lines to read.  Negative values
indicate that one should read up to the end of the connection.  The default
value is 1.}

\item{\dots}{passed on to the \code{readLines} function.}
}
\value{
The line reading iterator.
}
\description{
Returns an iterator over the lines of text from a connection.  It is a
wrapper around the standard \code{readLines} function.
}
\details{
Originally from the \code{iterators} package.
}
\examples{

# create an iterator over the lines of COPYING
it <- ireadLines(file.path(R.home(), "COPYING"))
nextOr(it)
nextOr(it)
nextOr(it)

}
\seealso{
\code{\link[base]{readLines}}
}
