% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recycle.R
\name{i_recycle}
\alias{i_recycle}
\title{Create a recycling iterator}
\usage{
i_recycle(iterable, times = NA_integer_, ...)
}
\arguments{
\item{iterable}{The iterable to recycle.}

\item{times}{integer.  Number of times to recycle the values in the}

\item{...}{Further arguments to be passed to
iterator.  Default value of \code{NA_integer_} means to recycle forever.}
}
\description{
Create an iterator that recycles a specified iterable. On the first
repeat the iterable is buffered into memory until it
finishes, then we repeat the same sequence of values.
}
\details{
Originally from the \code{itertools} package.
}
\examples{

# Recycle over 'a', 'b', and 'c' three times
i <- i_recycle(letters[1:3], 3)
as.character(i)

it <- i_recycle(1:3)
nextOr(it, NA) # 1
nextOr(it, NA) # 2
nextOr(it, NA) # 3
nextOr(it, NA) # 1
nextOr(it, NA) # 2
nextOr(it, NA) # 3
nextOr(it, NA) # 1

it2 <- i_recycle(1:3, times=2)
as.list(it2)

# Can return the results from a function.
it3 <- iteror(function() rnorm(1), count=Inf)
nextOr(it, NA)
nextOr(it, NA)
nextOr(it, NA)
nextOr(it, NA)

}
\keyword{utilities}
