% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irep.R
\name{i_rep}
\alias{i_rep}
\title{Create a repeating iterator}
\usage{
i_rep(iterable, times = NULL, length.out = NULL, each = NULL, ...)
}
\arguments{
\item{iterable}{The iterable to iterate over repeatedly.}

\item{times}{A vector giving the number of times to repeat each element if
the length is greater than one, or to repeat all the elements if the length
is one.  This behavior is less strict than \code{rep} since the length of an
iterable isn't generally known.}

\item{length.out}{non-negative integer.  The desired length of the output
iterator.}

\item{each}{non-negative integer.  Each element of the iterable is repeated
\code{each} times.}

\item{...}{passed along to \code{iteror(iterable, ...)}}
}
\description{
Create an iterator version of the \code{rep} function.
}
\details{
Originally from the \code{itertools} package.
}
\examples{

unlist(as.list(i_rep(1:4, 2)))
unlist(as.list(i_rep(1:4, each=2)))
unlist(as.list(i_rep(1:4, c(2,2,2,2))))
unlist(as.list(i_rep(1:4, c(2,1,2,1))))
unlist(as.list(i_rep(1:4, each=2, len=4)))
unlist(as.list(i_rep(1:4, each=2, len=10)))
unlist(as.list(i_rep(1:4, each=2, times=3)))

}
\seealso{
\code{\link[base]{rep}}
}
\keyword{utilities}
