% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_repeat.R
\name{i_repeat}
\alias{i_repeat}
\title{Create a repeating iterator}
\usage{
i_repeat(x, times)
}
\arguments{
\item{x}{The value to return repeatedly.}

\item{times}{The number of times to repeat the value.  Default value is
infinity.}
}
\value{
an \link{iteror}.
}
\description{
Create an iterator that returns a value a specified number of times.
}
\details{
Originally from the \code{itertools} package.
}
\examples{

# Repeat a value 10 times
unlist(as.list(i_repeat(42, 10)))

}
