% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iapply.R
\name{iapply}
\alias{iapply}
\title{Apply a function to each element of an iterator.}
\usage{
iapply(obj, f, ...)
}
\arguments{
\item{obj}{an iterable.}

\item{f}{a function}

\item{...}{Additional arguments will be passed along to \code{f}}
}
\value{
An iteror.
}
\description{
\code{iapply(obj, f)} returns the iteror that applies \code{f} to
each element of the given iterable \code{obj}. It is an iterator
equivalent of \code{lapply}.
}
\seealso{
To apply a function of multiple arguments to multiple
iterators, see \link{imap}. To split an array over margins (like
\code{iterators::iapply} use \code{\link[=iteror.array]{iteror(obj, by=MARGIN}}
}
\keyword{utilities}
