% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i_concat.R, R/ichain.R
\name{icollapse}
\alias{icollapse}
\alias{ichain}
\title{Iteror that chains multiple arguments together into a single iterator}
\usage{
icollapse(obj, ...)

ichain(...)
}
\arguments{
\item{obj}{an iterable.}

\item{...}{multiple iterable arguments}
}
\value{
iteror that iterates through each argument in sequence
}
\description{
\code{icollapse(obj)} takes an iterable that returns
iterables, and chains together all inner values of iterables into
one iterator. Analogous to \code{unlist(recursive=FALSE)}.

\code{ichain} for iterators is analogous to \code{\link[=c]{c()}} on vectors. \code{ichain}
constructs an \link{iteror} that returns elements from the first
argument until it is exhausted, then elements from the next
argument, and so on until all arguments have been exhausted.
}
\examples{
it <- ichain(1:3, 4:5, 6)
as.list(it)

it2 <- ichain(1:3, levels(iris$Species))
as.list(it2)
}
\author{
Peter Meilstrup
}
