% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icount.r
\name{icount}
\alias{icount}
\title{Iterator of numeric sequence with initial value and step size}
\usage{
icount(start = 0, step = 1, stop = NULL)
}
\arguments{
\item{start}{sequence's initial value}

\item{step}{sequence's step size}

\item{stop}{optional stopping point to sequence}
}
\value{
sequence's iteror
}
\description{
Constructs an \link{iteror} that generates a sequence of evenly spaced values
starting with \code{icount}. The step size is given by \code{step}. By
default, the sequence is neverending unless the optional \code{stop} is
provided.
}
\details{
NOTE: Use a negative \code{step} size to generate decreasing sequences.

Often used as an argument to \code{\link[iterors]{imap}} to
generate consecutive data points.
}
\examples{
it <- icount()
nextOr(it, NULL)
nextOr(it, NULL)
nextOr(it, NULL)

it2 <- icount(start=5.5, step=1.5)
nextOr(it2, NULL)
nextOr(it2, NULL)
nextOr(it2, NULL)

it3 <- icount(start=1, stop=3)
nextOr(it3, NULL)
nextOr(it3, NULL)
nextOr(it3, NULL)

}
