% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{idiv}
\alias{idiv}
\title{Dividing Iterator}
\usage{
idiv(n, ..., chunks, chunkSize)
}
\arguments{
\item{n}{number of times that the iterator will fire.  If not specified, it
will count forever.}

\item{\dots}{unused.}

\item{chunks}{the number of pieces that \code{n} should be divided into.
This is useful when you know the number of pieces that you want.  If
specified, then \code{chunkSize} should not be.}

\item{chunkSize}{the maximum size of the pieces that \code{n} should be
divided into.  This is useful when you know the size of the pieces that you
want.  If specified, then \code{chunks} should not be.}
}
\value{
The dividing iterator.
}
\description{
Returns an iterator that returns pieces of numeric value.
}
\details{
Originally from the `iterators` package.
}
\examples{

# divide the value 10 into 3 pieces
it <- idiv(10, chunks = 3)
nextOr(it)
nextOr(it)
nextOr(it)
nextOr(it, NULL)  # expect NULL

# divide the value 10 into pieces no larger than 3
it <- idiv(10, chunkSize = 3)
nextOr(it)
nextOr(it)
nextOr(it)
nextOr(it)
nextOr(it, NULL)  # end of iterator

}
\keyword{utilities}
