% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilimit.R
\name{ilimit}
\alias{ilimit}
\title{Create a limited iterator}
\usage{
ilimit(iterable, n)
}
\arguments{
\item{iterable}{Iterable to iterate over.}

\item{n}{Maximum number of values to return.}
}
\description{
Create an iterator that wraps a specified iterable a limited number of
times.
}
\examples{

# Limit icount to only return three values
as.list(ilimit(iterators::icount(), 3))

}
\keyword{utilities}
