% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipairwise.r
\name{ipairwise}
\alias{ipairwise}
\title{Iterator that returns elements of an object in pairs}
\usage{
ipairwise(object)
}
\arguments{
\item{object}{an iterable object}
}
\value{
an iterator that returns pairwise elements
}
\description{
Constructs an iterator of an iterable \code{object} that returns its elements
in pairs.
}
\examples{
it <- ipairwise(iterators::iter(letters[1:4]))
nextOr(it, NA) # list("a", "b")
nextOr(it, NA) # list("b", "c")
nextOr(it, NA) # list("c", "d")

it2 <- ipairwise(1:5)
nextOr(it2, NA) # list(1, 2)
nextOr(it2, NA) # list(2, 3)
nextOr(it2, NA) # list(3, 4)
nextOr(it2, NA) # list(4, 5)

}
