% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recycle.R
\name{irecycle}
\alias{irecycle}
\title{Create a recycling iterator}
\usage{
irecycle(iterable, times = NA_integer_)
}
\arguments{
\item{iterable}{The iterable to recycle.}

\item{times}{integer.  Number of times to recycle the values in the
iterator.  Default value of \code{NA_integer_} means to recycle forever.}
}
\description{
Create an iterator that recycles a specified iterable.
}
\details{
Originally from the \code{itertools} package.
}
\examples{

# Recycle over 'a', 'b', and 'c' three times
recycle(letters[1:3], 3)

}
\keyword{utilities}
