% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iunique.r
\name{irle}
\alias{irle}
\alias{irleinv}
\title{Run-length encoding iterator.}
\usage{
irle(obj, cmp = identical, ...)

irleinv(obj, ...)
}
\arguments{
\item{obj}{An iterable}

\item{cmp}{A function to use for comparison. It should take two
arguments and return \code{TRUE} or \code{FALSE}.}

\item{...}{further arguments forwarded to \code{\link[=iteror]{iteror(obj, ...)}}.}
}
\value{
An iterator returning entries of the form \code{list(length=n, value=X)}.

\code{irleinv} recreates the original data from the output of \code{irle}.
}
\description{
This is an iterator equivalent of \link{rle}; it produces one output
value for each run if identical values in its input, along with the
lenght of the run. \code{irle_inverse()} performs the inverse
transformstion.
}
\examples{
it <- isample(c(TRUE, FALSE), 1, replace=TRUE)
rle <- irle(it)
x <- take(rle, 10)
as.logical(irleinv(x))
}
\seealso{
idedupe
}
\author{
Peter Meilstrup
}
