% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{irnorm}
\alias{irnorm}
\alias{irunif}
\alias{irbinom}
\alias{irnbinom}
\alias{irpois}
\title{Random Number Iterators}
\usage{
irnorm(..., count)
}
\arguments{
\item{\dots}{arguments to pass to the underlying \code{rnorm} function.}

\item{count}{number of times that the iterator will fire.  If not specified,
it will fire values forever.}
}
\value{
An iterator that is a wrapper around the corresponding random number
generator function.
}
\description{
These function returns an iterators that return random numbers of various
distributions.  Each one is a wrapper around a standard \code{R} function.
}
\examples{

# create an iterator that returns three random numbers
it <- irnorm(1, count = 3)
nextOr(it)
nextOr(it)
nextOr(it)
nextOr(it, NULL)  # expect a StopIteration exception

}
\keyword{utilities}
