% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_iteror.r
\name{is.iteror}
\alias{is.iteror}
\alias{is.iterator}
\title{\code{is.iteror} indicates if an object is an iterator.}
\usage{
is.iteror(x)
}
\arguments{
\item{x}{any object.}
}
\description{
\code{is.iteror} indicates if an object is an iterator.
}
\examples{

it <- iteror(1:3)
stopifnot(is.iteror(it))
repeat {
  print(nextOr(it, break))
}

}
\keyword{utilities}
