% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{is_iteror}
\alias{is_iteror}
\title{Helper function that determines whether is an iteror object}
\usage{
is_iteror(object)
}
\arguments{
\item{object}{an R object}
}
\value{
logical value indicating whether \code{object} is of class
\code{iter}
}
\description{
Returns \code{TRUE} if the \code{object} is an object of class \code{iteror},
and \code{FALSE} otherwise.
}
