% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isplitIndices.R
\name{isplitIndices}
\alias{isplitIndices}
\title{Create an iterator of indices}
\usage{
isplitIndices(n, ...)
}
\arguments{
\item{n}{Maximum index to generate.}

\item{\dots}{Passed as the second and subsequent arguments to \code{idiv}
function.  Currently, \code{idiv} accepts either a value for \code{chunks}
or \code{chunkSize}.}
}
\value{
An iterator that returns vectors of indices from 1 to \code{n}.
}
\description{
Create an iterator of chunks of indices from 1 to \code{n}. You can specify
either the number of pieces, using the \code{chunks} argument, or the
maximum size of the pieces, using the \code{chunkSize} argument.
}
\details{
Originally from the \code{itertools} package.
}
\examples{

# Return indices from 1 to 17 in vectors no longer than five
it <- ihasNext(isplitIndices(17, chunkSize=5))
repeat print(nextOr(it, break))

# Return indices from 1 to 7 in four vectors
it <- ihasNext(isplitIndices(7, chunks=4))
repeat print(nextOr(it, break))

}
\seealso{
\code{\link[iterators]{idiv}, \link{isplitVector}}
}
\keyword{utilities}
