% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isplitRows.R
\name{isplitRows}
\alias{isplitRows}
\title{Create an iterator that splits a matrix into block rows}
\usage{
isplitRows(x, ...)
}
\arguments{
\item{x}{Matrix to iterate over.}

\item{\dots}{Passed as the second and subsequent arguments to \code{idiv}
function.  Currently, \code{idiv} accepts either a value for \code{chunks}
or \code{chunkSize}.}
}
\value{
An iterator that returns submatrices of \code{x}.
}
\description{
Create an iterator that splits a matrix into block rows. You can specify
either the number of blocks, using the \code{chunks} argument, or the
maximum size of the blocks, using the \code{chunkSize} argument.
}
\details{
Originally from the \code{itertools} package.
}
\examples{

# Split a matrix into submatrices with a maximum of three rows
x <- matrix(1:100, 10)
it <- isplitRows(x, chunkSize=3)
repeat print(nextOr(it, break))

# Split the same matrix into five submatrices
it <- isplitRows(x, chunks=5)
repeat print(nextOr(it, break))

}
\seealso{
\code{\link[iterators]{idiv}, \link{isplitCols}}
}
\keyword{utilities}
