% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itee.r
\name{itee}
\alias{itee}
\title{(Deprecated) Returns a list of n independent iterators from a single iterable object}
\usage{
itee(object, n = 2)
}
\arguments{
\item{object}{an iterable object}

\item{n}{the number of iterables to return}
}
\value{
a list of \code{n} iterators
}
\description{
(PBM) Note that this relies on \url{iter_deepcopy} which only works for
a subset of iterators, as described in the note on that page.
}
\details{
Constructs a list of \code{n} iterators, each of which iterates through an
iterable \code{object}.

If the \code{object} is an iterator (i.e., inherits from class \code{iter}),
\code{n} deep copies of \code{object} are returned. Otherwise, \code{object}
is passed to \code{\link[iterators]{iter}} \code{n} times.
}
\examples{
# Creates a list of three iterators.
# Each iterator iterates through 1:5 independently.
iter_list <- itee(1:5, n=3)

# Consumes the first iterator
unlist(as.list(iter_list[[1]])) == 1:5

# We can iterate through the remaining two iterators in any order.
nextOr(iter_list[[2]], NA) # 1
nextOr(iter_list[[2]], NA) # 2

nextOr(iter_list[[3]], NA) # 1
nextOr(iter_list[[3]], NA) # 2

nextOr(iter_list[[2]], NA) # 3
nextOr(iter_list[[2]], NA) # 4
nextOr(iter_list[[2]], NA) # 5

nextOr(iter_list[[3]], NA) # 3
nextOr(iter_list[[3]], NA) # 4
nextOr(iter_list[[3]], NA) # 5
}
