% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterators.R
\name{iter}
\alias{iter}
\alias{iter.default}
\alias{iter.iter}
\alias{iter.matrix}
\alias{iter.data.frame}
\alias{iter.function}
\title{Iterator Factory Functions}
\usage{
iter(obj, ...)
}
\arguments{
\item{obj}{an object from which to generate an iterator.}

\item{\dots}{additional arguments affecting the iterator.}

\item{by}{how to iterate.}

\item{chunksize}{the number of elements of \code{by} to return with each
call to \code{nextElem}.}

\item{checkFunc}{a function which, when passed an iterator value, return
\code{TRUE} or \code{FALSE}.  If \code{FALSE}, the value is skipped in the
iteration.}

\item{recycle}{a boolean describing whether the iterator should reset after
running through all it's values.}
}
\value{
The iterator.
}
\description{
\code{iter} is a generic function used to create iterator objects.
}
\examples{

# a vector iterator
i1 <- iter(1:3)
nextElem(i1)
nextElem(i1)
nextElem(i1)

# a vector iterator with a checkFunc
i1 <- iter(1:3, checkFunc = function(i) i\%\%2 == 0)
nextElem(i1)

# a data frame iterator by column
i2 <- iter(data.frame(x = 1:3, y = 10, z = c("a", "b", "c")))
nextElem(i2)
nextElem(i2)
nextElem(i2)

# a data frame iterator by row
i3 <- iter(data.frame(x = 1:3, y = 10), by = "row")
nextElem(i3)
nextElem(i3)
nextElem(i3)

# a function iterator
i4 <- iter(function() rnorm(1))
nextElem(i4)
nextElem(i4)
nextElem(i4)

}
\keyword{methods}
