% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{iter_deepcopy}
\alias{iter_deepcopy}
\title{Performs a deep copy of an iterator.}
\usage{
iter_deepcopy(iterator)
}
\arguments{
\item{iterator}{an iterator object that inherits from class 'iteror'}
}
\value{
a new iterator with its own state
}
\description{
This function only works on iterors constructed by
\url{iteror.default}, i.e. iterors created from a constant vector,
which have a "state" field. It will not work with name of the
higher order iterators created by the "i*()" functions in this
package.
}
