% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteror_array.R
\name{iteror.array}
\alias{iteror.array}
\title{Constructor method}
\usage{
\method{iteror}{array}(
  obj,
  ...,
  by = c("cell", "row", "column"),
  chunkSize,
  chunks,
  recycle = FALSE,
  drop = FALSE,
  rowMajor = TRUE
)
}
\arguments{
\item{by}{for \code{iteror.array}, how to split up an array or data frame. Can be "cell",
"row", "column", or numeric dimensions.}

\item{chunkSize}{the number of elements of \code{by} to return with
each call to \code{nextOr}.}

\item{drop}{Whether to drop the array dimensions enumerated over.}

\item{rowMajor}{If TRUE, will return slices in order with the first
indices varying fastest (same as in \link{i_enumerate}).}
}
\value{
an iteror yielding from \code{obj} along the specified dimensions.
}
\description{
Constructor method
}
\examples{
l <- iteror(letters, chunkSize=7)
as.list(l)

a <- array(1:8, c(2, 2, 2))

# iterate over all the slices
it <- iteror(a, by=3)
as.list(it)

# iterate over all the columns of each slice
it <- iteror(a, by=c(2, 3))
as.list(it)

# iterate over all the rows of each slice
it <- iteror(a, by=c(1, 3))
as.list(it)

}
