\name{itertools-package}
\alias{itertools-package}
\alias{itertools}
\docType{package}
\title{
The itertools Package
}
\description{
Functions for creating various iterators.  Many of the functions
are patterned after functions of the same name in the Python
itertools module, including \code{chain}, \code{product},
\code{izip}, \code{ifilter}, etc.
}
\details{
This package contains a collection of utility functions for
creating various kinds of iterators.

For a complete list of functions with individual help pages,
use \code{library(help="itertools")}.
}
\keyword{package}
