% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeout.R
\name{itimeout}
\alias{itimeout}
\title{Create a timeout iterator}
\usage{
itimeout(iterable, time)
}
\arguments{
\item{iterable}{Iterable to iterate over.}

\item{time}{The time interval to iterate for, in seconds.}
}
\description{
Create an iterator that iterates over another iterator for a specified
period of time, and then stops.  This can be useful when you want to search
for something, or run a test for awhile, and then stop.
}
\details{
Originally from the \code{itertools} package.
}
\examples{

# See how high we can count in a tenth of a second
length(as.list(itimeout(icount(), 0.1)))

}
\keyword{utilities}
