% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iwindow.R
\name{iwindow}
\alias{iwindow}
\title{Construct a sliding window over an iterator}
\usage{
iwindow(obj, n, tails = FALSE)
}
\arguments{
\item{obj}{An iterable.}

\item{n}{The width of the window to apply}

\item{tails}{Whether to include tails at the beginning and end. The
tails will be filled with NULL.}
}
\value{
an iteror.
}
\description{
Each element returned by \code{iwindow(obj)} consists of \code{n} consecutive
elements from the underlying \code{obj}, with the window advancing
forward by one element each iteration.
}
\examples{

#' @examples
it <- iwindow(iteror(letters[1:4]), 2)
nextOr(it, NA) # list("a", "b")
nextOr(it, NA) # list("b", "c")
nextOr(it, NA) # list("c", "d")

it2 <- iwindow(icount(5), 2)
nextOr(it2, NA) # list(1, 2)
nextOr(it2, NA) # list(2, 3)
nextOr(it2, NA) # list(3, 4)
nextOr(it2, NA) # list(4, 5)

it <- iwindow(letters[1:4], 2)
nextOr(it, NA) # list("a", "b")
nextOr(it, NA) # list("b", "c")
nextOr(it, NA) # list("c", "d")

it <- iwindow(letters[1:4], 3)
nextOr(it) # list("a", "b", "c")
nextOr(it) # list("b", "c", "d")

it <- iwindow(letters[1:4], 3, tails=TRUE)
nextOr(it) # list(NULL, NULL, "a")
nextOr(it) # list(NULL, "a", "b")
nextOr(it) # list("a", "b", "c")
nextOr(it) # list("b", "c", "d")
nextOr(it) # list("c", "d", NULL)
nextOr(it) # list("d", NULL, NULL)
}
\author{
Peter Meilstrup
}
