% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/izip.R
\name{izip}
\alias{izip}
\title{Create an iterator over multiple iterables}
\usage{
izip(...)
}
\arguments{
\item{\dots}{The iterables to iterate over.}
}
\description{
Create an iterator that iterates over multiple iterables, returning the
values as a list.
}
\details{
Originally from the `itertools` package.
}
\examples{

# Iterate over two iterables of different sizes
as.list(izip(a=1:2, b=letters[1:3]))

}
\keyword{utilities}
