% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterators.R
\name{nextElem}
\alias{nextElem}
\alias{nextElem.containeriter}
\alias{nextElem.funiter}
\title{Get Next Element of Iterator}
\usage{
nextElem(obj, ...)
}
\arguments{
\item{obj}{an iterator object.}

\item{\dots}{additional arguments that are ignored.}
}
\value{
The value.
}
\description{
\code{nextElem} is a generic function used to produce values. If a
\code{checkFunc} was specified to the constructor, the potential iterated
values will be passed to the \code{checkFunc} until the \code{checkFunc}
returns \code{TRUE}. When the iterator has no more values, it calls stop
with the message 'StopIteration'.
}
\examples{

it <- iteror(c("a", "b", "c"))
print(nextElem(it))
print(nextElem(it))
print(nextElem(it))

}
\keyword{methods}
