% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteror.R
\name{nextElemOr}
\alias{nextElemOr}
\title{Retreive the next element from an iteror.}
\usage{
nextElemOr(obj, or, ...)
}
\arguments{
\item{obj}{An [iteror]}

\item{or}{If the iterator has reached its end, an argument that
will be forced and returned}
}
\description{
Retreive the next element from an iteror.
}
