% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consume.r
\name{nth}
\alias{nth}
\title{Returns the nth item of an iteror}
\usage{
nth(it, n, or)
}
\arguments{
\item{n}{The location of the desired element to return}

\item{or}{An argument to force and return if the iteror is consumed.}

\item{iteror}{an iteror object}
}
\value{
The nth element of the iteror or the result of forcing \code{or}
}
\description{
Returns the \code{n}th item of an \code{iteror} after advancing the
iteror \code{n} steps ahead. If the \code{iteror} is entirely consumed,
the \code{default} value is returned instead. That is, if either \code{n >
length(iteror)} or \code{n} is 0, then the \code{iteror} is consumed.
}
\examples{
it <- iteror(1:10)
# Returns 5
nth(it, 5, NA)

it2 <- iteror(letters)
# Returns 'e'
nth(it2, 5, NA)

it3 <- iteror(letters)
# Returns default value of NA
nth(it3, 42, NA)

it4 <- iteror(letters)
# Returns default value of "foo"
nth(it4, 42, or="foo")

}
