% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py.R
\name{py_iteror}
\alias{py_iteror}
\title{Wrap an iteror to use with Python code. Requires the
\code{reticulate} package.}
\usage{
py_iteror(obj, ...)
}
\arguments{
\item{obj}{An iterable object}

\item{...}{Passed along to \link{iteror}.}
}
\description{
Wrap an iteror to use with Python code. Requires the
\code{reticulate} package.
}
