% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icollect.R
\name{take}
\alias{take}
\title{Return the first n elements of an iterable object as a list}
\usage{
take(obj, n = 1, mode = "list", ...)
}
\arguments{
\item{obj}{an iterable object}

\item{n}{the number of elements to return.}

\item{mode}{What mode to use for the output vector.}

\item{...}{passed along to \link{iteror} constructor.}
}
\value{
a list of the first \code{n} items of the iterable
\code{obj}
}
\description{
Returns the first \code{n} elements of an iterable \code{object} as a list.
If \code{n} is larger than the number of elements in \code{object}, the
entire iterator is consumed.
}
\details{
A function \code{take} first appeared in package \code{itertools2}.
It is basically an alias for \link{collect} but defaults to taking
one element.
}
\examples{
take(1:10, 3) # 1 2 3
take(icount(), 10) # 1:10
take(icount(5), 10) # 1 2 3 4 5
}
