% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writedf.R
\name{writedf.combiner}
\alias{writedf.combiner}
\title{Create an object that contains a combiner function}
\usage{
writedf.combiner(filenames)
}
\arguments{
\item{filenames}{Names of files to write column data to.}
}
\description{
Create an object that contains a combiner function.
}
\keyword{utilities}
