\name{writedf.combiner}
\alias{writedf.combiner}
\title{Create an object that contains a combiner function}
\description{
Create an object that contains a combiner function.
}
\usage{
writedf.combiner(filenames)
}
\arguments{
  \item{filenames}{Names of files to write column data to.}
}
\keyword{utilities}
