% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consume.r
\name{consume}
\alias{consume}
\title{Consumes the first n elements of an iterator}
\usage{
consume(iteror, n = 0)
}
\arguments{
\item{iteror}{an iteror object}

\item{n}{The number of elements to consume.}
}
\value{
Nothing, i.e., \code{invisible(NULL)}
}
\description{
Advances the iterator n-steps ahead without returning anything.
}
\details{
If \code{n} is 0, the iterator is consumed entirely. Similarly, if \code{n}
is larger than the length of the iterator, the iterator is consumed entirely.
}
\examples{
it <- iteror(1:10)
# Skips the first 5 elements
consume(it, n=5)
# Returns 6
nextOr(it, NA)

it2 <- iteror(letters)
# Skips the first 4 elements
consume(it2, 4)
# Returns 'e'
nextOr(it2, NA)

}
