% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilimit.R
\name{i_limit}
\alias{i_limit}
\title{Limit the length of an iterator.}
\usage{
i_limit(iterable, n, ...)
}
\arguments{
\item{iterable}{Iterable to iterate over.}

\item{n}{Maximum number of values to return.}

\item{...}{Extra arguments for \code{iteror(iterable, ...)}}
}
\value{
an \link{iteror} which will stop after yielding \code{n} values.
}
\description{
Create an iterator that limits the specified iterable to a specified number of items.
}
\details{
Originally from the \code{itertools} package.
}
\examples{

# Limit icount to only return three values
as.list(i_limit(icount(), 3))

}
