% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iapply.R
\name{iapply}
\alias{iapply}
\title{Apply a function to each element of an iterator.}
\usage{
iapply(obj, f, ...)
}
\arguments{
\item{obj}{an iterable.}

\item{f}{a function}

\item{...}{Other arguments that will be passed along to \code{f}}
}
\value{
An iteror.
}
\description{
\code{iapply(obj, f)} returns the iteror that applies \code{f} to
each element of the given iterable \code{obj}. It is an iterator
equivalent of \code{lapply}.
}
\seealso{
\code{\link{imap}} To applying a function to parallel iterables in parallel. \link{iteror.array}

The \code{iterators} package included an \code{iapply} to iterate over the
margins of an array; for equivalent behavior you can use
\verb{iteror(ARRAY, by=MARGIN)]}
}
\keyword{utilities}
