% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ichunk.R
\name{ichunk}
\alias{ichunk}
\title{Create a chunking iterator}
\usage{
ichunk(iterable, chunkSize, mode = "list")
}
\arguments{
\item{iterable}{Iterable to iterate over.}

\item{chunkSize}{Maximum number of values from \code{iterable} to return in
each value issued by the resulting iterator.}

\item{mode}{Mode of the objects returned by the iterator.}
}
\description{
Create an iterator that issues lists of values from the underlying iterable.
This is useful for manually \dQuote{chunking} values from an iterable.
}
\details{
Originally from the `itertools` package.
}
\examples{

# Split the vector 1:10 into "chunks" with a maximum length of three
it <- ichunk(1:10, 3)
repeat print(unlist(nextOr(it, break)))

# Same as previous, but return integer vectors rather than lists
it <- ichunk(1:10, 3, mode='integer')
repeat print(unlist(nextOr(it, break)))

}
\seealso{
\code{\link{isplitVector}}
}
\keyword{utilities}
