% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{icount}
\alias{icount}
\alias{icountn}
\title{Counting Iterators}
\usage{
icount(count)
}
\arguments{
\item{count}{number of times that the iterator will fire.  If not specified,
it will count forever.}

\item{vn}{vector of counts.}
}
\value{
The counting iterator.
}
\description{
Returns an iterator that counts starting from one.
}
\details{
Originally from the `iterators` package.
}
\examples{

# create an iterator that counts from 1 to 3.
it <- icount(3)
nextOr(it)
nextOr(it)
nextOr(it)
nextOr(it, NULL)  # expect NULL

}
\keyword{utilities}
