% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra.R
\name{icount}
\alias{icount}
\alias{icountn}
\title{Counting Iterators}
\usage{
icount(count, recycle = FALSE)
}
\arguments{
\item{count}{number of times that the iterator will fire.  If not specified,
it will count forever.}

\item{recycle}{Whether to restart the count after finishing.}
}
\value{
The counting iterator.
}
\description{
Returns an iterator that counts starting from one.
}
\details{
Originally from the \code{iterators} package.
}
\examples{

# create an iterator that counts from 1 to 3.
it <- icount(3)
nextOr(it)
nextOr(it)
nextOr(it)
nextOr(it, NULL)  # expect NULL

}
\seealso{
For more control over starting number and step size, see \link{iseq}.
}
\keyword{utilities}
