% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ienumerate.r
\name{ienumerate}
\alias{ienumerate}
\alias{ienum}
\alias{ienumerate.array}
\title{Iterator that returns the elements of an object along with their indices}
\usage{
ienumerate(obj, ...)

ienum(obj, ...)

\method{ienumerate}{array}(
  obj,
  ...,
  by = c("cell", "row", "column"),
  chunksize = 1L,
  recycle = FALSE,
  drop = FALSE
)
}
\arguments{
\item{obj}{object to return indefinitely.}

\item{...}{Undocumented.}

\item{by}{Which margind to split an array by.}

\item{chunkSize}{How large a chunk to take along the specified
dimension.}

\item{recycle}{Whether to restart the iterator after finishing the
array.}

\item{drop}{Whether to drop marginalized dimensions. Will have no
effect if chunkSize > 1.}
}
\value{
iterator that returns the values of \code{obj} along with the
index of the object.
}
\description{
Constructs an iterator that returns the elements of an object along with each
element's indices. Enumeration is useful when looping through an
\code{object} and a counter is required.

The \code{ienumerate} method for arrays allows splitting an
array by arbitrary margins, and by more than one margin. The
\code{index} element returned will be a vector (or if \code{chunkSize} > 1, a
matrix) of indices.
}
\details{
This function is intended to follow the convention used in Python's
\code{enumerate} function where the primary difference is that a list is
returned instead of Python's \code{tuple} construct.

Each call to \code{\link[iterators]{nextElem}} returns a list with two
elements:
\describe{
\item{index:}{a counter}
\item{value:}{the current value of \code{object}}
}

\code{ienum} is an alias to \code{ienumerate} to save a few keystrokes.

First appeared in package \code{iterators2}.
}
\examples{
set.seed(42)
it <- ienumerate(rnorm(5))
as.list(it)

# Iterates through the columns of the iris data.frame
it2 <- ienum(iris)
nextOr(it2, NA)
nextOr(it2, NA)
nextOr(it2, NA)
nextOr(it2, NA)
nextOr(it2, NA)

a <- array(1:27, c(3, 3, 3))
as.list(ienumerate(a, by=c(1, 2), chunkSize=2, drop=TRUE))
as.list(ienumerate(a, chunkSize=7))
}
\author{
Peter Meilstrup
}
