% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifilter.R
\name{ifilter}
\alias{ifilter}
\alias{ifilterfalse}
\title{Create a filtering iterator}
\usage{
ifilter(pred, iterable)
}
\arguments{
\item{pred}{A function that takes one argument and returns \code{TRUE} or
\code{FALSE}.}

\item{iterable}{The iterable to iterate over.}
}
\description{
The \code{ifilter} and \code{ifilterfalse} functions create iterators that
return a subset of the values of the specified iterable. \code{ifilter}
returns the values for which the \code{pred} function returns \code{TRUE},
and \code{ifilterfalse} returns the values for which the \code{pred}
function returns \code{FALSE}.
}
\details{
Originally from the `itertools` package.
}
\examples{

# Return the odd numbers between 1 and 10
as.list(ifilter(function(x) x \%\% 2 == 1, icount(10)))

# Return the even numbers between 1 and 10
as.list(ifilterfalse(function(x) x \%\% 2 == 1, icount(10)))

}
\keyword{utilities}
