% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihasNext.R
\name{ihasNext}
\alias{ihasNext}
\title{Create an iterator that supports the hasNext method}
\usage{
ihasNext(iterable)
}
\arguments{
\item{iterable}{an iterable object, which could be an iterator.}
}
\value{
An \code{ihasNext} iterator that wraps the specified iterator and
supports the \code{hasNext} method.
}
\description{
\code{ihasNext} is a generic function that indicates if the iterator has
another element.
}
\examples{

  it <- ihasNext(c('a', 'b', 'c'))
  while (hasNext(it))
    print(nextElem(it))

}
\keyword{utilities}
