% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recycle.R
\name{irecycle}
\alias{irecycle}
\title{Create a recycling iterator}
\usage{
irecycle(iterable, times = NA_integer_)
}
\arguments{
\item{iterable}{The iterable to recycle. If it is given a bare function (this behavior is copied from \code{icycle}}

\item{times}{integer.  Number of times to recycle the values in the
iterator.  Default value of \code{NA_integer_} means to recycle forever.}
}
\description{
Create an iterator that recycles a specified iterable.
}
\details{
Originally from the \code{itertools} package.

it <- irecycle(1:3)
nextOr(it, NA) # 1
nextOr(it, NA) # 2
nextOr(it, NA) # 3
nextOr(it, NA) # 1
nextOr(it, NA) # 2
nextOr(it, NA) # 3
nextOr(it, NA) # 1

it2 <- irecycle(1:3, times=2)
as.list(it2)
}
\section{Can return the results from a function.}{
it3 <- irecycle(function() rnorm(1))
nextOr(it, NA)
nextOr(it, NA)
nextOr(it, NA)
nextOr(it, NA)
}

\examples{

# Recycle over 'a', 'b', and 'c' three times
recycle(letters[1:3], 3)

}
\keyword{utilities}
