% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ireduce.R
\name{ireduce}
\alias{ireduce}
\alias{iaccum}
\alias{sum.iteror}
\alias{prod.iteror}
\title{Compute the sum, product, or general reduction of an iterator.}
\usage{
ireduce(obj, fun = `+`, init = 0, ...)

iaccum(obj, fun = `+`, init = 0, ...)

\method{sum}{iteror}(..., na.rm)

\method{prod}{iteror}(..., na.rm)
}
\arguments{
\item{obj}{an iterable object}

\item{fun}{A function of as least two arguments.}

\item{init}{A starting value.}

\item{...}{Extra parameters to be passed to \code{f}}

\item{na.rm}{Whether to drop NA values when computing sum or prod.}
}
\value{
The result of accumulation.
}
\description{
Compute the sum, product, or general reduction of an iterator.

\code{iaccum(obj)} returns the iterator containing
each intermediate result. The default settings
produce a cumulative sum.
}
\examples{
it <- icount(5)
ireduce(it, `+`) # sum(1:5)

it <- icount(5)
ireduce(it, paste0, "") # "12345"

it <- icount(5)
ireduce(it, `*`, 1) # prod(1:5)

}
\author{
Peter Meilstrup
}
