% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_iteror.R
\name{is.iteror}
\alias{is.iteror}
\alias{is.iterator}
\title{\code{is.iteror} indicates if an object is an iteror.}
\usage{
is.iteror(x)
}
\arguments{
\item{x}{any object.}
}
\value{
TRUE if the object has class \code{iteror}.
}
\description{
\code{is.iteror} indicates if an object is an iteror.
}
\examples{

it <- iteror(1:3)
stopifnot(is.iteror(it))
repeat {
  print(nextOr(it, break))
}

}
