% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itripletwise.r
\name{itripletwise}
\alias{itripletwise}
\title{Iterator that returns adjacent triplets from an iterator}
\usage{
itripletwise(object)
}
\arguments{
\item{object}{an iterable object}
}
\value{
an iterator that returns tripletwise elements
}
\description{
Constructs an iterator of an iterable \code{object} that returns its elements
in a sliding window three elements wide.
}
\examples{
it <- itripletwise(iteror(letters[1:4]))
nextOr(it) # list("a", "b", "c")
nextOr(it) # list("b", "c", "d")

it2 <- itripletwise(1:5)
nextOr(it2) # list(1, 2, 3)
nextOr(it2) # list(2, 3, 4)
nextOr(it2) # list(3, 4, 5)

}
