% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_iteror.r
\name{nextOr}
\alias{nextOr}
\title{Retreive the next element from an iteror.}
\usage{
nextOr(obj, or, ...)
}
\arguments{
\item{obj}{An \link{iteror}}

\item{or}{If the iterator has reached its end, this argument
will be forced and returned.}

\item{...}{Other arguments.}
}
\description{
Retreive the next element from an iteror.
}
