% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_iteror.R
\name{nextOr}
\alias{nextOr}
\title{Retreive the next element from an iteror.}
\usage{
nextOr(obj, or, ...)
}
\arguments{
\item{obj}{An \link{iteror}.}

\item{or}{If the iterator has reached its end, this argument
will be forced and returned.}

\item{...}{Other arguments may be used by specific iterors.}
}
\value{
Either the next value of \code{iteror}, or the value of \code{or}.
}
\description{
Retreive the next element from an iteror.
}
