% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py.R
\name{r_to_py.iteror}
\alias{r_to_py.iteror}
\title{Wrap an iteror to pass to Python functions, via the
\code{reticulate} package.}
\usage{
\method{r_to_py}{iteror}(x, convert = FALSE, ...)
}
\arguments{
\item{x}{An iterable object.}

\item{...}{Passed along to \link{iteror(x, ...)}.}
}
\value{
a Python iterator.
}
\description{
Wrap an iteror to pass to Python functions, via the
\code{reticulate} package.
}
\examples{
pit <- r_to_py(iseq(2, 11, 5))
reticulate::iter_next(pit, NULL)
reticulate::iter_next(pit, NULL)
reticulate::iter_next(pit, NULL)

triangulars <- icount() |> i_accum() |> i_limit(10)
builtins <- reticulate::import_builtins()
builtins$sum(triangulars) # r_to_py is called automatically
}
